remove tags that were used only once.
I sometimes look at the overview of tags with `jrnl --tags` and there
I found it annoying to see all the tags that appeared only once. Those
are now filtered out and I print a comment saying that they were filtered out

Also changed the order (sorted by counts) in which the tags are printed. For
a CLI application I think it is nicer to see the more important version at the end