Added ability to list accessible journals.
Previously, if a user forgot the exact name of a journal, they could check the configuration file to retrieve this information. This patch adds support for performing this task using the jrnl interface.

The newly added '-ls' option retrieves the available journals from the jrnl configuration files and displays each journal name on a separate line.

This patch also updates the jrnl help display and the usage documentation to reflect these changes.