editor: use shell to launch editor to allow spaces in editor command
If I want to set a filetype in vim on opening the journla, e.g.:

  vim -c 'set ft=notes' -c Goyo -f

the set and ft=notes will be split and interpreted as two files in vim.
Using the shell to execute the command makes it possible to have editor
commands with quoted/escaped spaces in them.