Add ability to search in entry title and body
jrnl was missing ability to search in entry titles and message bodies.

This commit adds a '-search TEXT' option which, if specified, causes
jrnl to display only those entries which have 'TEXT' in either their
title or message body.

Specifying the '-search' option also implies that the compose mode
won't be used.

Fixes #384.