Update Poetry requirements for testing latest Python version (#898)
Poetry errors out if we try to run a different version of Python that
what is specified in the config file. We want to run tests against the
latest Python version so we can see problems before they happen. This
fix updates the config file so Poetry will allow us to test. Nothing is
committed after the update.