Extensive modifications to overview.md (#957)
* Extensive modifications to overview.md.

I tried to add clarity and details while maintaining the spirit of the original
document. However, it might be a bit too 'dry' now. I'd be happy to liven it up
a bit. I'm only serious when I feel like I _have_ to be.

One of my opinions (which may be at odds with yours) is that the documentation
should emphasize `jrnl`'s advantages without downplaying any other existing
solutions.

If I have time, I'd like to add more information about the documented benefits
of journaling, particularly the mental health aspects. That will probably need
its own page, but I'm new here, so I don't want to overstay my welcome. :)

* More changes to overview.md in response to feedback

"why keep a journal" section removed -- it could be re-added as its own page,
but it's not front-page material

* More changes to overview.md:
- fixed up the headings
- added information about multi-platform support

This is a work in progress and _not_ ready for prime-time.
- need to add to Command-Line Interface section, among other things

* More changes to overview.md:

- moved "`jrnl` is a simple..." to "Command-Line Interface" section