Change styling of terminal on docs site, small copy changes (#1038)
The terminal now overlaps the header (the purple area at the top of the
page) and the main area (the white area underneath). This draws more
focus to it, and quite literally puts the terminal front and center.

This also fixes a few typos, and updates the commands in the terminal to
match jrnl v2.5 updates.