Change approach for docs workflow to use pa11y-ci
The previous plugin looked nice with the comments and annotations, but
apparently doesn't work for PRs. It can only run on the pr target (our
branch), which means that the tests are never actually run against the
new code.

As you can guess, this is pretty useless for PRs. So, this PR fixes
that, and uses the very nice pa11y-ci (instead of the normal pa11y).
While it doesn't upload annotations, the output is very readable and
seems like the best we're going to get.