Fix version header detection in changelog
This updates the regex to find the version header in our changelog.
There was a bug that affected patch, alpha, and beta versions. For
example, if we were looking for "v2.5" we might find "v2.5.1" or
"v2.5.1-alpha" or other versions like that.

This led to the changelog not being cleared properly on update, and old
versions sticking around longer than they should.