Fixes to changelog generator (#1127)
* Fix version header detection in changelog

This updates the regex to find the version header in our changelog.
There was a bug that affected patch, alpha, and beta versions. For
example, if we were looking for "v2.5" we might find "v2.5.1" or
"v2.5.1-alpha" or other versions like that.

This led to the changelog not being cleared properly on update, and old
versions sticking around longer than they should.

* re-add option to include author link for merged PRs in changelog

This option was accidentally left out when migrating the changelog
generator tool to github actions.

* fix conditional in changelog workflow

In github actions, booleans always end up as strings. So, it's not
enough to test `env.FULL_RELEASE` because both 'true' and 'false' will
evaluate as true. This fixes the conditional to check for the string
value 'true' instead of relying on a boolean.